chore: use `NDataTable` for table list
